+++ src/CMakeLists.txt
@@ -44,14 +44,14 @@ set(GUI_SOURCES
 )
 
 set(LOG_SOURCES
-    ${CMAKE_CURRENT_SOURCE_DIR}/log/qgis_debug.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/log/qgis_debug.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/log/QgisDebug.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/log/QgisDebug.h
 )
 
 set(OPENGL_SOURCES
-    ${CMAKE_CURRENT_SOURCE_DIR}/opengl/camera.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/opengl/Camera.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/opengl/Primitive.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/opengl/camera.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/opengl/Camera.h
     ${CMAKE_CURRENT_SOURCE_DIR}/opengl/Primitive.h
 )
 
@@ -64,9 +64,9 @@ set(SCHOOLCORE_SOUCRES
 
 set(STANDALONE_SOURCES
     main.cpp
-    mainwindow.cpp
-    mainwindow.h
-    mainwindow.ui
+    MainWindow.cpp
+    MainWindow.h
+    MainWindow.ui
 )
 
 if(BUILD_AS_PLUGIN)
@@ -86,19 +86,19 @@ if(BUILD_AS_PLUGIN)
 else()
 
     if(${QT_VERSION_MAJOR} GREATER_EQUAL 6)
-        qt_add_executable(transfer_school
+        qt_add_executable(3Dschool
             MANUAL_FINALIZATION
             ${SCHOOLCORE_SOUCRES}
             ${STANDALONE_SOURCES}
             ${RESOURCES}
         )
     # Define target properties for Android with Qt 6 as:
-    #    set_property(TARGET transfer_school APPEND PROPERTY QT_ANDROID_PACKAGE_SOURCE_DIR
+    #    set_property(TARGET 3Dschool APPEND PROPERTY QT_ANDROID_PACKAGE_SOURCE_DIR
     #                 ${CMAKE_CURRENT_SOURCE_DIR}/android)
     # For more information, see https://doc.qt.io/qt-6/qt-add-executable.html#target-creation
     else()
         if(ANDROID)
-            add_library(transfer_school SHARED
+            add_library(3Dschool SHARED
                 ${SCHOOLCORE_SOUCRES}
                 ${STANDALONE_SOURCES}
                 ${RESOURCES}
@@ -106,7 +106,7 @@ else()
     # Define properties for Android with Qt 5 after find_package() calls as:
     #    set(ANDROID_PACKAGE_SOURCE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/android")
         else()
-            add_executable(transfer_school
+            add_executable(3Dschool
                 ${SCHOOLCORE_SOUCRES}
                 ${STANDALONE_SOURCES}
                 ${RESOURCES}
@@ -114,15 +114,15 @@ else()
         endif()
     endif()
 
-    target_link_libraries(transfer_school PRIVATE ${LIBS})
+    target_link_libraries(3Dschool PRIVATE ${LIBS})
 
     # Qt for iOS sets MACOSX_BUNDLE_GUI_IDENTIFIER automatically since Qt 6.1.
     # If you are developing for iOS or macOS you should consider setting an
     # explicit, fixed bundle identifier manually though.
     if(${QT_VERSION} VERSION_LESS 6.1.0)
-      set(BUNDLE_ID_OPTION MACOSX_BUNDLE_GUI_IDENTIFIER com.example.transfer_school)
+      set(BUNDLE_ID_OPTION MACOSX_BUNDLE_GUI_IDENTIFIER com.example.3Dschool)
     endif()
-    set_target_properties(transfer_school PROPERTIES
+    set_target_properties(3Dschool PROPERTIES
         ${BUNDLE_ID_OPTION}
         MACOSX_BUNDLE_BUNDLE_VERSION ${PROJECT_VERSION}
         MACOSX_BUNDLE_SHORT_VERSION_STRING ${PROJECT_VERSION_MAJOR}.${PROJECT_VERSION_MINOR}
@@ -131,13 +131,13 @@ else()
     )
 
     include(GNUInstallDirs)
-    install(TARGETS transfer_school
+    install(TARGETS 3Dschool
         BUNDLE DESTINATION .
         LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
         RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
     )
 
     if(QT_VERSION_MAJOR EQUAL 6)
-        qt_finalize_executable(transfer_school)
+        qt_finalize_executable(3Dschool)
     endif()
 endif()
