+++ src/gui/StyleManager.h
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef STYLEMANAGER_H
+#define STYLEMANAGER_H
 #include <QString>
 
 class StyleManager {
@@ -6,3 +7,5 @@ public:
     static void initializeStyle();
     static QString loadQSS(const QString& path);
 };
+
+#endif // STYLEMANAGER_H
