+++ src/log/QgisDebug.h
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef QGISDEBUG_H
+#define QGISDEBUG_H
 #include <QString>
 #include <qdebug.h>
 //#include <qfiledialog.h>
@@ -10,3 +11,5 @@
 #include <QTextStream>
 
 void logMessage(const QString &message,Qgis::MessageLevel level = Qgis::MessageLevel::Info);
+
+#endif // QGISDEBUG_H
